backend/requirements.txt
django==5.1.5
psycopg2-binary>=2.9
django-cors-headers>=3.14.0
djangorestframework>=3.14.0
djangorestframework-simplejwt